Добавил привилегии администраторов и права доступа (reverted from commit 01306325f8d66da45205cfcdd341aa84ed73b7be)

+++ www/admin/actions/delete-admin.php
@@ -6,7 +6,12 @@ $action = $_GET["action"];
 if(isset($action)) {
     switch ($action) {
         case 'delete':
+        if ($_SESSION['admin_login'] == 'admin') {
             $delete = mysql_query("DELETE FROM admins WHERE id = '$id'",$link);
+            }
+            else {
+                $msgerror = 'У вас нет прав на удаление администраторов!';
+            }
             break;
             
         default:
